+++ src/converse-core.js
@@ -70,6 +70,7 @@
         'converse-core',
         'converse-disco',
         'converse-dragresize',
+        'converse-fullscreen',
         'converse-headline',
         'converse-mam',
         'converse-minimize',
@@ -80,6 +81,7 @@
         'converse-register',
         'converse-roomslist',
         'converse-rosterview',
+        'converse-singleton',
         'converse-vcard'
     ];
 
@@ -325,6 +327,7 @@
             storage: 'session',
             strict_plugin_dependencies: false,
             synchronize_availability: true,
+            view_mode: 'overlayed', // Choices are 'overlayed', 'fullscreen', 'mobile'
             websocket_url: undefined,
             whitelisted_plugins: [],
             xhr_custom_status: false,